Moved all router hooks to hooks/router.js
