+++ web/src/hooks/redux.js
@@ -4,7 +4,7 @@ import { useSelector } from 'react-redux';
 import { useParams } from 'react-router-dom';
 import { keys, sortBy } from 'lodash';
 
-import { useCustomDatetime } from '../helpers/router';
+import { useCustomDatetime } from './router';
 
 export function useCurrentZoneHistory() {
   const { zoneId } = useParams();
